requirements.txt
python-dotenv
langchain-core
langchain-groq
streamlit 
watchdog